feat(custom-forms): show custom forms in tables in app styles
