+++ src/data-workspace/custom-form/index.js
@@ -1,0 +1,1 @@
+export { CustomForm } from './custom-form.js'
