+++ src/data-workspace/entry-form.js
@@ -2,7 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { Button, InputField } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { CustomForm } from './custom-form.js'
+import { CustomForm } from './custom-form/index.js'
 import { DefaultForm } from './default-form.js'
 import styles from './entry-form.module.css'
 import { SectionForm } from './section-form/index.js'
